refactor : MyPage Component 분리 #PAR-380
