+++ feature/my_page/src/main/java/online/partyrun/partyrunapplication/feature/my_page/component/MyPageProfile.kt
@@ -1,0 +1,106 @@
+package online.partyrun.partyrunapplication.feature.my_page.component
+
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.zIndex
+import online.partyrun.partyrunapplication.core.designsystem.component.RenderAsyncUrlImage
+import online.partyrun.partyrunapplication.core.designsystem.icon.PartyRunIcons
+import online.partyrun.partyrunapplication.feature.my_page.R
+
+@Composable
+fun ProfileContent(
+    navigateToProfile: () -> Unit,
+    userName: String,
+    userProfile: String,
+) {
+    Column(
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        ProfileHeader(
+            navigateToProfile = navigateToProfile,
+            userName = userName
+        )
+        ProfileImage(
+            userProfile = userProfile
+        )
+    }
+}
+
+@Composable
+private fun ProfileHeader(
+    navigateToProfile: () -> Unit,
+    userName: String
+) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(top = 20.dp, bottom = 10.dp),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(
+            text = userName,
+            style = MaterialTheme.typography.titleLarge,
+            color = MaterialTheme.colorScheme.onPrimary,
+            modifier = Modifier
+                .align(Alignment.Center)
+                .zIndex(1f) // 이미지와 동일한 Z-축 위치에 배치
+        )
+
+        IconButton(
+            onClick = {
+                navigateToProfile()
+            },
+            modifier = Modifier
+                .size(45.dp)
+                .padding(end = 10.dp)
+                .align(Alignment.CenterEnd)
+        ) {
+            Icon(
+                painter = painterResource(id = PartyRunIcons.edit),
+                tint = MaterialTheme.colorScheme.onPrimary,
+                contentDescription = stringResource(id = R.string.edit_desc)
+            )
+        }
+    }
+}
+
+@Composable
+private fun ProfileImage(
+    userProfile: String
+) {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = Modifier
+            .size(80.dp)
+            .aspectRatio(1f, matchHeightConstraintsFirst = true)
+            .clip(CircleShape)
+            .border(3.dp, MaterialTheme.colorScheme.onPrimary, CircleShape)
+            .zIndex(1f)
+    ) {
+        RenderAsyncUrlImage(
+            modifier = Modifier.fillMaxSize(),
+            imageUrl = userProfile,
+            contentDescription = null
+        )
+    }
+}
